+++ templates/python.flask/{{cookiecutter.project_slug}}/ci/at.sh
@@ -1,3 +1,5 @@
 #!/bin/sh
 
-echo AT
+IMAGE=$(sirius docker_image_name | head -n 1)
+
+sirius docker_deploy:{{cookiecutter.project_slug}},${IMAGE},server=DEPLOY_SERVER,ports="9201;8080",env="ENV\=prd"