+++ templates/python.flask/{{cookiecutter.project_slug}}/config/default.py
@@ -1,1 +1,4 @@
+HTTP_HOST = ''
+HTTP_PORT = 8080
+
 LOG="/var/{{cookiecutter.project_slug}}/error.log"