+++ templates/python.flask/{{cookiecutter.project_slug}}/run.py
@@ -1,6 +1,9 @@
 """deimos entry point
 
 """
+import gevent.monkey
+gevent.monkey.patch_all()
+
 import logging
 import os
 import os.path