+++ templates/python.flask/{{cookiecutter.project_slug}}/{{cookiecutter.project_slug}}/views/__init__.py
@@ -1,0 +1,3 @@
+from faq import version
+
+__all__ = ['version']