+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change log
 
+## [0.0.4]
+
+### Added
+- Add release docker image python.flask
+- Add faq and version for python.flask
+
 ## [0.0.2]
+
 ### Added
 - Add create, version, list, deploy sub-command